Move route tests to Integration folder
